requirements.txt
langchain
openai
chromadb
streamlit
beautifulsoup4
requests
python-dotenv